More changes in how encryption works

+++ application/controllers/InstancedataController.php
@@ -204,21 +204,21 @@ class InstancedataController extends QFrame_Controller_Admin {
     $file = $_FILES['instanceFile']['tmp_name'];
     $filename = $_FILES['instanceFile']['name'];
     
-    if (preg_match('/\.enc$/i', $filename)) {
+    if (preg_match('/\.enc\.zip$/i', $filename)) { // Encrypted zip
       if (!isset($decryptID) || $decryptID == 0) {
         throw new Exception('Key not specified for encrypted file');
       }
       $crypto = new CryptoModel(array('cryptoID' => $decryptID));
-      if (preg_match('/\.zip\.enc$/i', $filename)) {
+      if (preg_match('/\.xml.enc.zip$/i', $filename)) { // Encrypted zip consisting of just the xml
+        $decrypted = $crypto->decrypt(file_get_contents($file), 'instance-responses.xml');
+        $import = $decrypted;
+      }
+      elseif (preg_match('/\.enc\.zip$/i', $filename)) { // Encrypted zip consisting of xml and attachments
         $decrypted = $crypto->decrypt(file_get_contents($file), 'instance-responses.zip');
         $tempfile = tempnam(PROJECT_PATH . DIRECTORY_SEPARATOR . 'tmp', 'zip');
         unlink($tempfile);
         file_put_contents($tempfile, $decrypted);
         $import = new ZipArchiveModel(null, array('filename' => $tempfile));
-      }
-      elseif (preg_match('/\.xml.enc$/i', $filename)) {
-        $decrypted = $crypto->decrypt(file_get_contents($file), 'instance-responses.xml');
-        $import = $decrypted;
       }
       else {
         throw new Exception('Unrecognized file extension [' . $filename . ']');
